pyproject.toml
[project]
name = "jolto-pipeline"
version = "0.1.0"
description = "CGI jewellery advertising video pipeline: brief + product image -> short ad clip"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.7",
    "typer>=0.12",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "fal-client>=0.5",
    "google-genai>=0.3",
    "pillow>=10.3",
    "open-clip-torch>=2.24",
    "torch>=2.2",
    "rich>=13.7",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "python-multipart>=0.0.9",
]

[project.scripts]
jolto = "pipeline.run:app"
jolto-web = "server.main:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["pipeline*", "server*"]

[tool.setuptools.package-data]
"server" = ["web/*"]
